pyproject.toml
[project]
name = "mcp-server-flomo-python"
version = "0.1.0"
description = "MCP Server for Flomo integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "anyio>=4.5.0"
]
